refactor - extract responsibility of checking negative number and parse to int
- NumberExtractor 객체의 extractNumbers 메서드의 책임을 따로 다른 클래스(PositiveIntNumber)로 추출했습니다(기존: 클래스 내부 메서드)
